Rename `side-by-side` to `aside` and `mini-card` to `card`
This commit renames `side-by-side` to `aside` and `mini-card` to `card`
to resolve a bug with the module.exports syntax. The only reason we were
using that syntax was to allow for dashes in the markdoc tag names. By
simplifying the tag names we can switch to the simpler export syntax.

Additionally, this now allows us to upgrade to Next.js 14.1.0 which
has a bug in it related to the module.exports syntax. Finally, this
commit also removes some dead code related to the tags.